+++ qml/components/GameCard.qml
@@ -237,6 +237,7 @@ FocusScope {
     }
 
     Column {
+        objectName: "cardCaption"
         anchors.top: cover.bottom
         anchors.topMargin: 10
         anchors.left: parent.left
@@ -307,15 +308,7 @@ FocusScope {
                 font.weight: Font.DemiBold
             }
         }
-        ProtonDbBadge {
-            objectName: "cardProtonDbBadge"
-            width: parent.width
-            gameSource: root.gameSource
-            appId: root.appId
-            font.pixelSize: 10
-            elide: Text.ElideRight
-            fetchEnabled: root.visible
-        }
+
     }
 
     MouseArea {